店铺注册dao层与service层实现&测试

+++ src/main/java/com/cheng/o2o/entity/Area.java
@@ -13,7 +13,7 @@ public class Area {
     /**
      * ID
      */
-    private Integer areaId;
+    private Long areaId;
     /**
      * 名称
      */
@@ -31,11 +31,11 @@ public class Area {
      */
     private Date lastEditTime;
 
-    public Integer getAreaId() {
+    public Long getAreaId() {
         return areaId;
     }
 
-    public void setAreaId(Integer areaId) {
+    public void setAreaId(Long areaId) {
         this.areaId = areaId;
     }
